Move object and block properties inside 'keymesh' PropertyGroup (#7)
Registers Keymesh properties on objects and blocks inside `PropertyGroup`. Properties aren't exposed to users anymore and clutter Custom Properties panel, and avoid unintentional breaks. Word "Keymesh" is removed from properties except for objects "Keymesh Data", which is needed to be displayed in animation editors on fcurve.

Versioning for transferring properties in old files, as well as entire animation is added as well.