Use random number in new_object_id
Previous method checked for lowest unused ID in the same object type as active object. That method is good for producing manageable numbers (and I might revisit later), but it caused problems when appending/linking, since lower numbers like 1 or 2 can already exist in file it's being appended it. That caused Keymesh data to break.

While this method doesn't extinguish that problem (append/link handlers are needed for that) it minimizes the risks of that happening.

Versioning is also modified to assign new random ID instead of reusing old one.