Handle duplicates in "Convert to Separate Objects" operator (#10)
For both 3D printing and rendering workflows it's good to track duplicates and handle them in some way. Duplicate is a block which was used more than once in Keymesh animation, and therefore (before this patch) is created multiple times as new object. Detecting and removing duplicates can avoid printing unnecessary replacement parts, or animating more objects than need to be.

- Adds properties for handling duplicates and doesn't create objects if "Handle Duplicates" is selected
- Instancing method of duplicate handling for Rendering workflow: still creates new objects, but instances data from match, instead of copying
- Reuse method same object by animating it on every frame Keymesh block was appearing on for Rendering workflow
- Print how many instances were created, or which objects were reused for animation
- For 3D printing workflow delete duplicates and print about blocks that were used on multiple frames