Fix (unreported): Blocks can have same names
When name of the block is changed in UI it automatically changes name of the object data as well, but if user inputted name that was already used block name would get .001 added, while block didn't. This created all sorts of problems and allowed blocks to share same name.

Update function now checks if object data name has been changed to something else besides inputted name, and if it was, changed block name as well.