I18n: Add empty French translations
This file was auto-generated by running Manage UI Translations's
"Refresh I18n Data..." operator. It is not usable yet, since nothing
is translated and the file is not even used.

When we need to update the translations, e. g. if the UI has changed,
we can call the operator again and select the add-on.

Be careful to select only the languages you actually want to setup in
the "I18n Update Translation" panel, otherwise you'll get all of them.