I18n: Add po file for French
This GNU gettext po file is more human-readable than the add-on's .py
file, and it can be opened with most graphical translation
applications. This is the file that can be given to translators.

It is not necessary to actually commit .po file, as they can be
imported back into the .py file (see later commit).

The export was done automatically using the "Export PO..." operator.