Remove BUILD tool
While keeping this script in add-on repositories has it's upsides, it has prooved to be inconvenient when working on multiple add-ons alone, as it needs constant updates between Blender versions in multiple repositories.

So I decided to remove tool from repos and maintain it's single copy separately. In future I might make this (and other tools) in their own separate repo, which can be included as submodule in add-on repos.