Fix (unreported): Keymesh f-curve removed despite action slot being shared
This was happening in cases where we were removing f-curve on already non-Keymesh object (basically in `remove_keymesh_properties` function). `has_shared_action_slot` function was returning that the slot had 1 Keymesh user, therefore it was not shared, and was removed. But in reality slot was shared between Keymeesh and non-Keymesh objects and needed to stay exactly for that one Keymesh user.

Now the function returns True if action slot has at least one Keymesh user besides object itself.