Refactor convert operator, part III (Geometry Nodes workflow)
Adds a new conversion method to the Convert operator that converts Keymesh animation in a Geometry Nodes modifier workflow. New duplicate objects is created for each Keymesh object, as with the other methods, but in this case modifiers and animation data are removed from them. They're only kept on the main duplicate object that holds the Geometry Nodes modifier. All the objects created from Keymesh blocks are imported in the modifier with "Object Info" nodes, and outputted one-by-one with "Index Switch", index of which is revealed as modifier input. In animated workflow that input is keyframed to recreate exact Keymesh animation as well.

This is probably the most faithful way of converting Keymesh, because the design is very similar. Keymesh object is also a holder that reveals object datas one-by-one. In complex cases this is also most non-destructive, because user will not have to maintain multiple sets of modifiers and actions for on each separate block. Just like with Keymesh animating & modifying one object will affect all of the object datas it outputs.

---

This commit was also only going to add a new conversion method, but for the third time it turned into a larger refactor, because more things became clear. Namely, it's not possible to support duplicate handling in new conversion methods. That means for them we always have to create one object per Keymesh block, and we don't care about how many times they were keyframed.

Duplicate handling was originally developed for the 3D printing workflow (which I'm sure nobody has used besides me), but it's also very useful in some advanced workflows, some of which were possible (like reporting duplicates, object-data instancing), some of which were planned (applying modifiers on each frame or keyframe). However, the best course of action for now would be to simply remove duplicate handling and above workflows. This will greatly simplify code and UI, as options between static and animated workflows will mostly be the same. Only difference between those two modes now is whether new objects visibility (or the property controlling it) is animated or not.

In near future duplicate handling will be back as "Advanced" workflow option. But at the moment it's better to simplify the code and design, and built it on top of the simple design, rather than further convolute it now.

Details:
- Renamed `animation_method` to `convert_method`. Renamed "Keyframe" method to "Simple", removed "None" method.
- Removed options for "Naming Convention", "Handle Duplicates", and "Handling Method"
- Separate "use_drivers" option is removed for static workflow, since it now uses same convert method as animated.

User-Facing Changes:
- Revamped UI for operator.
- New option for animated workflow that excludes blocks that were not used in animation (0 keyframes). It simply doesn't create objects for them.