Drop the versnioning
Versioning has been in the add-on for a long time, and has become a maintenance burden. Even though there might still be few cases that need versioning, I think it's time to drop it. Especialy because next version of Keymesh will be only for Blender 5.0 or later, and also might need to version newer stuff as well.